+++ src/lib.rs
@@ -1,14 +1,21 @@
-//! Define actions and their input binds and then see if its `pressing`, `pressed` or
-//! `released`. This library handles variable pressure through the `action_val` function aswellas
+#![cfg_attr(test, feature(test))]
+//! Define actions and their list of input binds. and then see if its `pressing`, `pressed` or
+//! `released`. This library handles variable pressure through the `value` function as well as
 //! multiple variable pressure inputs for things like 1d axis (through `axis`) and 2d vectors 
 //! (through `dir` and `dir_max_len_1`). This makes it easy to have things like 3d camera controls
-//! applied for both mouse movement and the gamepads right stick. The input map also can get the
-//! `mouse_pos`, what was `recently_pressed` (used for rebinding things) and the `text_typed`
-//! (useful for typing) to make sure it is fully featured. For better user control there are the
-//! `mouse_scale` and `scroll_scale` variables to make sensitivity align with eveything else
-//! 0-1 range and a `press_sensitivity` to control when a action counts as being pressed. Finaly,
-//! theres an input_map! macro to reduce boilerplate and increase readability.
-//! ```
+//! applied for both mouse movement and the gamepads right stick.
+//!
+//! The binds are structured by a list of `(Action, bind, bind, ..)`. An action is pressed if any
+//! of its binds are pressed. Binds are described as either `[InputCode, InputCode, ..]` or
+//! InputCode. A bind is pressed if all its InputCodes are pressed.
+//!
+//! The input map also can get the `mouse_pos`, what was `recently_pressed` (used for rebinding
+//! things) and the `text_typed` (useful for typing) to make sure it is fully featured.
+//! For better user control there are the `mouse_scale` and `scroll_scale` variables to make
+//! sensitivity align with eveything else 0-1 range and a `press_sensitivity` to control when an
+//! action counts as being pressed. Finaly, theres an `input_map!` macro to reduce boilerplate and
+//! increase readability.
+//! ```no_run
 //! use winit::{
 //!     window::*, application::*, keyboard::*,
 //!     event_loop::*, event::*
@@ -20,9 +27,9 @@
 //! enum Actions{ Foo }
 //! use Actions::*;
 //!
-//! let input = input_map!(
-//!     (Foo, KeyCode::Space, GamepadInput::South)
-//! );
+//! let input = { use base_input_codes::*; input_map!(
+//!     (Foo, KeyCode::Space, [GamepadInput::South, RightTrigger])
+//! ) };
 //! let ev = EventLoop::new().unwrap();
 //! let gilrs = Gilrs::new().unwrap();
 //! ev.run_app(&mut App { window: None, input, gilrs }).unwrap();
@@ -66,32 +73,170 @@ mod input;
 mod input_code;
 pub use crate::input::*;
 pub use crate::input_code::*;
-/// Creates new input map with inputed input codes bound to the acompaning action.
-/// Anything that impliments `into<InputCode>` can be bound to an action
+
+/// Outputs an input with the inputed binds.
+///
+/// The input is structured by a list of `(Action, bind, bind, ..)`. An action is pressed if any of
+/// its binds are pressed.
+///
+/// Binds are described as either `[InputCode, InputCode, ..]` or InputCode. A bind is pressed if all
+/// its InputCodes are pressed.
 /// ```
-/// use Action::*;
 /// use winit_input_map::*;
-/// use winit::{keyboard::KeyCode, event::MouseButton};
-/// #[derive(Hash, PartialEq, Eq, Clone, Copy)]
+/// use Action::*;
+/// #[derive(PartialEq, Eq, Clone, Copy, Hash, Debug)]
 /// enum Action {
-///     Jump,
-///     Left,
-///     Right,
-///     Interact
+///     Select, Undo, Redo, Confirm
 /// }
-/// let mut input = input_map!(
-///     (Jump,     KeyCode::Space                    ),
-///     (Left,     KeyCode::KeyA, KeyCode::ArrowLeft ),
-///     (Right,    KeyCode::KeyD, KeyCode::ArrowRight),
-///     (Interact, MouseButton::Left                 )
-/// );
+/// let input = { use base_input_codes::*; input_map!(
+///     (Select, MouseButton::Left, ShiftLeft, ShiftRight),
+///     (Action::Undo,  [KeyZ, ControlLeft], [KeyZ, ControlRight]),
+///     (Redo,  [KeyR, ControlLeft], [KeyR, ControlRight]),
+///     (Confirm, MouseButton::Left, Enter)
+/// ) };
 /// ```
 #[macro_export]
 macro_rules! input_map {
-    () => { InputMap::<()>::empty() };
-    ( $( ( $x:expr, $( $k:expr ),* ) ),* ) => {
-        InputMap::new(&[ $(
-            ($x, vec![ $( $k.into(), )* ]),
-        )*])
+    () => { $crate::InputMap::empty() };
+    ( $( ( $x: expr, $( $k: expr ),* ) ),* ) => {
+        $crate::InputMap::new(
+            &$crate::binds!($(
+                ($x, $( $k ),* )
+            ),*)
+        )
     };
+}
+/// Outputs binds in the expected format for `add_binds`, `set_binds` and `InputMap::new` though in
+/// the latter case `input_map!` should be used to skip the middle man.
+///
+/// The input is structured by a list of `(Action, bind, bind, ..)`. An action is pressed if any of
+/// its binds are pressed.
+///
+/// Binds are described as either `[InputCode, InputCode, ..]` or InputCode. A bind is pressed if all
+/// its InputCodes are pressed.
+/// ```
+/// use winit_input_map::*;
+/// use Action::*;
+/// #[derive(PartialEq, Eq, Clone, Copy, Hash, Debug)]
+/// enum Action {
+///     Select, Undo, Redo, Confirm
+/// }
+/// let mut input = input_map!((Select, base_input_codes::ShiftLeft));
+///
+/// let binds = { use base_input_codes::*; binds!(
+///     (Select, MouseButton::Left, ShiftRight),
+///     (Action::Undo,  [KeyZ, ControlLeft], [KeyZ, ControlRight]),
+///     (Redo,  [KeyR, ControlLeft], [KeyR, ControlRight]),
+///     (Confirm, MouseButton::Left, Enter)
+/// ) };
+///
+/// input.add_binds(&binds);
+/// ```
+#[macro_export]
+macro_rules! binds {
+    ( $( ( $x: expr, $( $k: expr ),* ) ),* ) => { {
+        $crate::as_binds(&[ $(
+                ($x, &[ $( (&$k as &dyn $crate::AsBind) ),* ])
+        ),* ])
+    } };
+}
+/// Takes in more flexible deffinition of binds, the `binds!` macro is however, still prefered.
+pub fn as_binds<'a, F: std::hash::Hash + Copy>(binds: &'a [(F, &'a [&'a dyn AsBind])]) -> crate::Binds<F> {
+    let mut result = Vec::new();
+    for (action, binds) in binds {
+        let mut sub_result = Vec::new();
+        for bind in *binds {
+            sub_result.push(bind.as_bind());
+        }
+        result.push((*action, sub_result));
+    }
+    result
+}
+
+trait AsInputCode { fn as_code(&self) -> InputCode; }
+impl<T: Into<InputCode> + Copy> AsInputCode for T { fn as_code(&self) -> InputCode { (*self).into() } }
+
+/// Dictates what can be turned into binds for the purposes of the `as_binds()` method and the
+/// `binds!()` and `input_map!()` macros
+pub trait AsBind { fn as_bind(&self) -> Vec<InputCode>; }
+impl<T: AsInputCode> AsBind for T { fn as_bind(&self) -> Vec<InputCode> { vec![self.as_code()] } }
+impl<T: AsInputCode, const X: usize> AsBind for [T; X] { 
+    fn as_bind(&self) -> Vec<InputCode> {
+        let mut result = Vec::new();
+        for i in self {
+            let bind: InputCode = i.as_code();
+            result.push(bind);
+        }
+        result
+    }
+}
+impl<T: AsInputCode> AsBind for [T] { 
+    fn as_bind(&self) -> Vec<InputCode> {
+        let mut result = Vec::new();
+        for i in self {
+            let bind: InputCode = i.as_code();
+            result.push(bind);
+        }
+        result
+    }
+}
+impl<T: AsInputCode> AsBind for &[T] { 
+fn as_bind(&self) -> Vec<InputCode> {
+    let mut result = Vec::new();
+    for i in *self {
+        let bind: InputCode = i.as_code();
+        result.push(bind);
+    }
+    result
+}
+}
+impl AsBind for Vec<&dyn AsInputCode> {
+    fn as_bind(&self) -> Vec<InputCode> {
+        let mut result = Vec::new();
+        for i in self {
+            let bind: InputCode = i.as_code();
+            result.push(bind);
+        }
+        result
+    }
+}
+impl AsBind for &[&dyn AsInputCode] {
+    fn as_bind(&self) -> Vec<InputCode> {
+        let mut result = Vec::new();
+        for i in *self {
+            let bind: InputCode = i.as_code();
+            result.push(bind);
+        }
+        result
+    }
+}
+impl AsBind for [&dyn AsInputCode] {
+    fn as_bind(&self) -> Vec<InputCode> {
+        let mut result = Vec::new();
+        for i in self {
+            let bind: InputCode = i.as_code();
+            result.push(bind);
+        }
+        result
+    }
+}
+impl<const X: usize> AsBind for &[&dyn AsInputCode; X] {
+    fn as_bind(&self) -> Vec<InputCode> {
+        let mut result = Vec::new();
+        for i in *self {
+            let bind: InputCode = i.as_code();
+            result.push(bind);
+        }
+        result
+    }
+}
+impl<const X: usize> AsBind for [&dyn AsInputCode; X] {
+    fn as_bind(&self) -> Vec<InputCode> {
+        let mut result = Vec::new();
+        for i in self {
+            let bind: InputCode = i.as_code();
+            result.push(bind);
+        }
+        result
+    }
 }